Tue, Jan 23, 2024, 12:24 PM -03:00

+++ calculator.c
@@ -6,6 +6,6 @@ int main(void)
     long x = get_long("X: ");
     long y = get_long("Y: ");
 
-    float z =(float) x / (float) y;
-    printf("%lf\n", z);
+    double z =(double) x / (double) y;
+    printf("%.20f\n", z);
 }
